requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
flask-cors==5.0.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
multidict==6.3.2
numpy
openai==0.28.0
opencv-python==4.11.0.86
propcache==0.3.1
pydantic==2.11.2
pydantic_core==2.33.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3

google-generativeai==0.4.0

google-genai==1.13.0

tensorflow
tensorflow_hub